refactor(T-1.12): feature-folders + lint rules
