+++ apps/web/src/features/repositories/constants.ts
@@ -1,0 +1,1 @@
+export const OPTIMISTIC_ID_PREFIX = "optimistic:";
