+++ apps/web/src/features/repositories/types.ts
@@ -1,0 +1,22 @@
+import type {
+  ConnectedRepo,
+  GithubRepo,
+} from "@commit-analyzer/contracts";
+
+export type GithubEnvelope = {
+  status: 200;
+  body: { items: GithubRepo[] };
+  headers: Headers;
+};
+
+export type ConnectedEnvelope = {
+  status: 200;
+  body: { items: ConnectedRepo[] };
+  headers: Headers;
+};
+
+export type RepositoriesPageData = {
+  userId: string;
+  initialGithub: GithubRepo[];
+  initialConnected: ConnectedRepo[];
+};
