+++ packages/eslint-config/next.js
@@ -1,9 +1,10 @@
 // Module boundary preset for apps/web (Next.js).
 //
 // Encodes the component inventory in docs/06-frontend.md §1–2 and the
-// golden rules in docs/02-architecture-global.md §7: feature folders are
-// sibling-isolated, ui primitives may not depend on features, and no web
-// code may reach into apps/api (ADR-0002 bounded contexts).
+// golden rules in docs/02-architecture-global.md §7: feature folders live
+// under src/features/* and are sibling-isolated (enforced by the local
+// `local-web/no-cross-feature-imports` rule), ui primitives may not depend
+// on features, and no web code may reach into apps/api (ADR-0002).
 
 import boundaries from "eslint-plugin-boundaries";
 import globals from "globals";
@@ -13,10 +14,13 @@ import base from "./base.js";
 const elements = [
   { type: "app-route", pattern: "src/app" },
   { type: "ui", pattern: "src/components/ui" },
+  { type: "layout-chrome", pattern: "src/components/layout" },
+  { type: "providers", pattern: "src/components/providers" },
   {
     type: "feature",
-    pattern: "src/components/*",
+    pattern: "src/features/*",
     capture: ["name"],
+    mode: "folder",
   },
   { type: "lib", pattern: "src/lib" },
   { type: "i18n", pattern: "src/i18n" },
@@ -48,6 +52,7 @@ export default [
       "boundaries/include": [
         "src/app/**/*.{ts,tsx}",
         "src/components/**/*.{ts,tsx}",
+        "src/features/**/*.{ts,tsx}",
         "src/lib/**/*.{ts,tsx}",
         "src/i18n/**/*.{ts,tsx}",
       ],
@@ -63,11 +68,34 @@ export default [
           rules: [
             {
               from: { type: "app-route" },
-              allow: { to: { type: ["feature", "ui", "lib", "i18n"] } },
+              allow: {
+                to: {
+                  type: [
+                    "feature",
+                    "ui",
+                    "layout-chrome",
+                    "providers",
+                    "lib",
+                    "i18n",
+                  ],
+                },
+              },
             },
             {
               from: { type: "feature" },
+              allow: {
+                to: { type: ["feature", "ui", "lib", "i18n"] },
+              },
+            },
+            {
+              from: { type: "providers" },
               allow: { to: { type: ["ui", "lib", "i18n"] } },
+            },
+            {
+              from: { type: "layout-chrome" },
+              allow: {
+                to: { type: ["ui", "lib", "i18n", "layout-chrome"] },
+              },
             },
             { from: { type: "ui" }, allow: { to: { type: "lib" } } },
             { from: { type: "i18n" }, allow: { to: { type: "lib" } } },
